move all feed tests into the same file
